fix: use Long for fileSize in MigrationV1ToV2 NewVideoState.Downloaded
VideoEntity.State.Downloaded.fileSizeはLongだが、Migration側のNewVideoState.Downloaded.fileSizeが
Intのままだったため、大きなファイルサイズでデシリアライズエラーになる可能性があった。

+++ datasource/src/main/kotlin/net/turtton/ytalarm/datasource/local/migration/MigrationV1ToV2.kt
@@ -127,7 +127,7 @@ object MigrationV1ToV2 : Migration(1, 2) {
             @SerialName("internal_link")
             val internalLink: String,
             @SerialName("file_size")
-            val fileSize: Int,
+            val fileSize: Long,
             @SerialName("is_streamable")
             val isStreamable: Boolean
         ) : NewVideoState
@@ -370,7 +370,7 @@ object MigrationV1ToV2 : Migration(1, 2) {
 
         is OldVideoState.Downloaded -> NewVideoState.Downloaded(
             internalLink = old.internalLink,
-            fileSize = old.fileSize,
+            fileSize = old.fileSize.toLong(),
             isStreamable = old.isStreamable
         )
     }
